+++ tests/Unit/Syntax/RangeTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace JeroenG\Explorer\Tests\Unit\QueryBuilders;
+namespace JeroenG\Explorer\Tests\Unit\Compound;
 
 use InvalidArgumentException;
 use JeroenG\Explorer\Domain\Syntax\Range;