+++ changelog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ## Added
-- Added the `newAggregate()` function to customize the aggregate query builder
+- Added the `newCompound()` function to customize the compound query
 
 ### Fixed
 - Boosts may now completely be left out